back/requirements.txt
Flask
python-dotenv
folium
geopy
requests
numpy
pandas
black
flake8
colorlog
colorama
shapely